requirements.txt
amqp==2.6.0
arabic-reshaper==2.1.1
argon2-cffi==20.1.0
billiard==3.6.3.0
cairocffi==1.1.0
CairoSVG==2.4.2
celery==4.4.6
certifi==2020.12.5
cffi==1.14.0
chardet==4.0.0
cssselect2==0.3.0
defusedxml==0.6.0
Django==2.2.13
django-celery-results==1.2.1
future==0.18.2
html5lib==1.1
idna==2.10
importlib-metadata==1.7.0
kombu==4.6.11
names==0.3.0
Pillow==7.1.2
pycparser==2.20
PyJWT==1.7.1
PyPDF2==1.26.0
Pyphen==0.9.5
python-bidi==0.4.2
python-dateutil==2.8.0
python-dotenv==0.15.0
pytz==2020.1
redis==3.5.3
reportlab==3.5.46
requests==2.25.1
six==1.15.0
sqlparse==0.3.1
tinycss2==1.0.2
twilio==6.51.0
urllib3==1.26.2
vine==1.3.0
WeasyPrint==52.2
webencodings==0.5.1
zipp==1.2.0
